@@ -4,7 +4,9 @@
   "private": true,
   "workspaces": [
     "vim-js",
-    "mrujs"
+    "mrujs",
+    "binaries",
+    "common"
   ],
   "main": "index.js",
   "author": "David",
@@ -24,7 +26,7 @@
     "lerna": "^3.22.1",
     "prettier": "^2.2.0",
     "ts-loader": "^8.0.11",
-    "typescript": "^4.1.2",
+    "typescript": "^4.2.4",
     "webpack": "^5.9.0",
     "webpack-cli": "^4.2.0"
   }
