@@ -1,6 +1,6 @@
 module.exports = {
   rules: {
-    "@typescript-eslint/no-var-requires": "off",
-    "no-console": "off",
+    '@typescript-eslint/no-var-requires': 'off',
+    'no-console': 'off',
   },
-};
+}
